LayoutTests/inspector/console/console-xml-document-expected.txt

CONSOLE MESSAGE: line 9: [object Document]
Tests that XML document contents are logged using the correct case in the console.

#document<MixedCase> Test </MixedCase> console-xml-document.html:9
